Relocate past Remark tests to a new location and added extra tests for special characters. Add tests for toStringRepresentation which has a different implementation from normal toString method
